refactor: 구매, 판매 기능 구현

+++ backend/src/main/java/com/welcome/tteoksang/game/dto/ProductTradeInfo.java
@@ -11,5 +11,5 @@ import lombok.*;
 public class ProductTradeInfo {
     private Integer productId;
     private Integer productQuantity;
-    private Integer productTotalCost;
+    private Long productTotalCost;
 }
